collapse ActiveSchema into Schema
