--- ruby/blueprints/active_schema.rb
@@ -1,9 +1,0 @@
-require_relative '../spaces/schema'
-require_relative 'schema'
-
-module Blueprints
-  class ActiveSchema < ::Spaces::Schema
-    extend Schema
-
-  end
-end
