+++ ruby/blueprints/schema.rb
@@ -1,3 +1,4 @@
+require_relative '../spaces/schema'
 require_relative '../frameworks/framework'
 require_relative '../repositories/repositories'
 require_relative '../sudos/sudos'
@@ -11,49 +12,51 @@ require_relative '../bindings/anchor'
 require_relative '../environments/environment'
 
 module Blueprints
-  module Schema
+  class Schema < ::Spaces::Schema
 
-    def outline
-      {
-        title: 0,
-        description: 0,
-        memory_usage: [0, { required: 0, recommended: 0 }],
-        protocol: 1,
-        licenses: [(1..), { label: 1, url: 1 }],
-        framework: 0,
-        repositories: 0,
-        sudos: 0,
-        os_packages: 0,
-        modules: 0,
-        packages: 0,
-        replacements: 0,
-        file_permissions: 0,
-        bindings: 0,
-        environment: 0
-      }
-    end
+    class << self
+      def outline
+        {
+          title: 0,
+          description: 0,
+          memory_usage: [0, { required: 0, recommended: 0 }],
+          protocol: 1,
+          licenses: [(1..), { label: 1, url: 1 }],
+          framework: 0,
+          repositories: 0,
+          sudos: 0,
+          os_packages: 0,
+          modules: 0,
+          packages: 0,
+          replacements: 0,
+          file_permissions: 0,
+          bindings: 0,
+          environment: 0
+        }
+      end
 
-    def collaborating_classes
-      [
-        Frameworks::Framework,
-        Repositories::Repositories,
-        Sudos::Sudos,
-        OsPackages::OsPackages,
-        Nodules::Nodules,
-        Packages::Packages,
-        Replacements::Replacement,
-        FilePermissions::FilePermissions,
-        Bindings::Bindings,
-        Bindings::Anchor,
-        Environments::Environment
-      ]
-    end
+      def collaborating_classes
+        [
+          Frameworks::Framework,
+          Repositories::Repositories,
+          Sudos::Sudos,
+          OsPackages::OsPackages,
+          Nodules::Nodules,
+          Packages::Packages,
+          Replacements::Replacement,
+          FilePermissions::FilePermissions,
+          Bindings::Bindings,
+          Bindings::Anchor,
+          Environments::Environment
+        ]
+      end
 
-    def outline_map
-      {
-        nodules: :modules,
-        anchor: :binding_anchor
-      }
+      def outline_map
+        {
+          nodules: :modules,
+          anchor: :binding_anchor
+        }
+      end
     end
 
   end
